Use common parts for both vue3 and 2 components
