+++ src/common/beforeCreate.js
@@ -1,0 +1,6 @@
+let idCounter = 0;
+
+export default function beforeCreate() {
+  this.idCounter = idCounter.toString();
+  idCounter += 1;
+}
