+++ src/common/data.js
@@ -1,0 +1,5 @@
+export default {
+  hovered: null,
+  isOpen: false,
+  hoveredIndex: -1,
+};
